Messages standardized following last group work session
In the last meeting several changes have been applied to the messages
in order to make them as regular as possible as well as consideration
about specific message data and about which parts of the original
protocol modify and which keep as is.

The GatewayMessage has not been tangled to try implementing a
generic gateway that only performs forwarding.

Message names have been changed to better represent what they
contains.

Nodes 6 and 7 are not tengled and will probably not be
implemented.

A first definition of the Networ enumeration has been given.

Signed-off-by: teo3300 <matteo.rogora@live.it>